Following test steps, some changes to library
https://rustwasm.github.io/book/game-of-life/testing.html
online tutorial seems to reference state of code prior to
last exercise from previous step:
The tutorial is still referencing Cell::Alive and Cell::Dead,
despite that model no longer being in use with the move to
one bit per cell described in the final exercise on the previous
page.
Had to... adapt the provided tutorial code with a version
that works with the 1-bit implementation.